requirements.txt
bs4
Pillow
cssutils
lxml
requests